added go rewrite of the project
