+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import (
+	"lib/jsonHandler"
+	"lib/redditHandler"
+)
+
